Enhance AI functionality for recipe analysis
- Refactored AI interface to support recipe-specific analysis methods for images and URLs.
- Updated OpenAI client to return structured recipe analysis results in JSON format.
- Modified RecipeService to handle AI analysis results and create recipes accordingly.
- Adjusted API routes to reflect new AI-powered recipe creation endpoints.
- Improved error handling for unsupported AI operations and added relevant error messages.
- Replaced test image with a new recipe image for better testing accuracy.

+++ internal/core/storage/mongo_storage_test.go
@@ -595,7 +595,7 @@ func TestCreateRecipeWithLargeDocument(t *testing.T) {
 	for i := 0; i < 100; i++ {
 		largeRecipe.Ingredients[i] = models.Ingredient{
 			Name:     fmt.Sprintf("Ingredient %d", i),
-			Quantity: i,
+			Quantity: float32(i),
 			Unit:     "unit",
 		}
 		largeRecipe.Steps[i] = fmt.Sprintf("Step %d: %s", i, strings.Repeat("Do something. ", 50))
